Semana 9 - Antes de realizar tarea#2
Funciona sin problemas - Hasta donde pude ver

+++ src/main/java/tienda/domain/Categoria.java
@@ -16,7 +16,7 @@ public class Categoria implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name="id_categoria")
-    private Integer idCategoria;
+    private Long idCategoria;
     
     @Column(unique=true, nullable=false, length=50)
     @NotNull
